Login page has signup button Newlisting redirects to listing page MergeDicts now in utils, imported to routes
