+++ app/templates/login.html
@@ -10,8 +10,11 @@
 	<p>
 	    {{ form.password.label }}<br>
 	    {{ form.password(size=32) }}
-        <p>{{ form.submit() }}</p>
-    </form>
+        <p>{{ form.submit() }}</p></form>
+        <p>Need an account? Register here: 
+          <form action = "/signup" method="GET" >  
+          <button type="submit" class="btn btn-sm btn-warning">Signup</button>
+        </form></p>
     <button onclick="myFunction()">Forgot Password?</button>
     <p id="lol"></p>
     <script>
